Relocate @composable codeblocks of InventoryView.kt into Dialogs.kt + insert Comments + fixed bugs
